Refactored project folders
