+++ TeamCode/src/main/java/hotcakes/BasicAuto.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package hotcakes;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
