+++ TeamCode/src/main/java/hotcakes/MotorControl.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package hotcakes;
 
 import org.firstinspires.ftc.robotcore.external.JavaUtil;
