+++ TeamCode/src/main/java/hotcakes/RobotHardware.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package hotcakes;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
